Create toys context
