+++ src/components/ToyFeatureForm.tsx
@@ -2,6 +2,7 @@ import { invoke } from "@tauri-apps/api";
 import { useEffect, useState } from "react";
 import Form from "react-bootstrap/Form";
 import { FeVCToyFeature } from "../../src-tauri/bindings/FeVCToyFeature";
+import { useToys } from "../context/ToysContext";
 import { ALTER_TOY, ALTER_TOY_DEBOUNCE } from "../data/constants";
 import { round0 } from "../utils";
 import "./ToyFeatureForm.css";
@@ -9,24 +10,24 @@ import "./ToyFeatureForm.css";
 type ToyFeatureFormProps = {
   toyId: number;
   feature: FeVCToyFeature;
-  refetchToys: () => void;
 };
 
-export default function ({ toyId, feature, refetchToys }: ToyFeatureFormProps) {
+export default function ({ toyId, feature }: ToyFeatureFormProps) {
   const [modifiedFeature, setModifiedFeature] =
     useState<FeVCToyFeature>(feature);
 
+  const { refetchToys } = useToys();
+
   useEffect(() => {
+    if (modifiedFeature == feature) return;
     async function setFeature() {
       await invoke(ALTER_TOY, { toyId: toyId, toyFeature: modifiedFeature });
     }
-    if (modifiedFeature != feature) {
-      const t = setTimeout(() => {
-        setFeature();
-        refetchToys();
-      }, ALTER_TOY_DEBOUNCE);
-      return () => clearTimeout(t);
-    }
+    const t = setTimeout(() => {
+      setFeature();
+      refetchToys();
+    }, ALTER_TOY_DEBOUNCE);
+    return () => clearTimeout(t);
   }, [modifiedFeature]);
 
   return (
