+++ src/components/Toys.css
@@ -1,0 +1,34 @@
+.toys-container {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  align-items: stretch;
+  font-size: 1rem;
+  min-width: 400px;
+  border-radius: var(--container-border-radius);
+  padding: 10px;
+  margin: 20px;
+  background-color: var(--toys-container);
+  box-shadow: 5px 5px 5px var(--toys-container-shadow);
+}
+
+.toy-container {
+  font-size: 2rem;
+  min-height: 100px;
+  border-radius: var(--container-border-radius);
+  margin: 10px;
+  background-color: var(--item);
+  box-shadow: 5px 5px 10px var(--item-shadow);
+  animation: enter-ani 250ms ease-out;
+}
+.toy {
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+}
+.toy-container .accordion {
+  padding-bottom: var(--container-border-radius);
+}
+.toy-container .accordion-body {
+  max-width: 360px;
+}
